Move { Header, Sidebar } to "src/layout"
컴포넌트가 점점 많아져서 폴더 구조를 개편하게 되었습니다

src
- components : 각 레이아웃안에 들어가게 될 요소들
- layout : 각 페이지의 레이아웃과 관련 (헤더, 사이드 바, 메인 등)
- pages : 라우팅을 통해 만나게 될 페이지들
